requirements.txt
django==2.2
requests
django-background-tasks